+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -11,6 +12,7 @@ class CommentController extends Controller
         $comment = new Comment();
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back()->with('info', 'A new comment is added');
@@ -18,8 +20,12 @@ class CommentController extends Controller
 
     public function delete($id) {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back()->with('info', 'A comment is deleted');
+        if (Gate::allows('comment-delete', $comment)) {
+            $comment->delete();
+            return back()->with('info', 'A comment is deleted');
+        } else {
+            return back()->with('info', 'Unauthorized');
+        }
     }
 }
